requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
lxml==5.3.0
packaging==24.1
pypdf==4.3.1
python-docx==1.1.2
typing_extensions==4.12.2
